+++ tools/eyeriss_search.py
@@ -168,8 +168,10 @@ def do_scheduling(args):
                      sw_solve_loopblocking=args.solve_loopblocking,
                      partition_hybrid=args.hybrid_partition,
                      partition_batch=args.batch_partition,
+                     partition_ifmaps=args.ifmaps_partition,
                      ntops=args.top,
-                     nprocesses=args.processes)
+                     nprocesses=args.processes,
+                     verbose=args.verbose)
 
     ## Search schedules.
 
@@ -213,8 +215,10 @@ def main(args):
     return 0
 
 
-if __name__ == '__main__':
-    ap = argparse.ArgumentParser()  # pylint: disable=invalid-name
+def argparser():
+    ''' Argument parser. '''
+
+    ap = argparse.ArgumentParser()
 
     ap.add_argument('net',
                     help='network name, should be a .py file under examples')
@@ -267,12 +271,22 @@ if __name__ == '__main__':
     ap.add_argument('--batch-partition', action='store_true',
                     help='Allow partitioning batch, i.e., consider data '
                          'parallelism.')
+    ap.add_argument('--ifmaps-partition', '--ifmap-partition',
+                    action='store_true',
+                    help='Allow partitioning ifmap channel dimension, which '
+                         'requires extra data synchronization.')
 
     ap.add_argument('-t', '--top', type=int, default=1,
                     help='Number of top schedules to keep during search.')
     ap.add_argument('-p', '--processes', type=int,
                     default=multiprocessing.cpu_count()/2,
                     help='Number of parallel processes to use for search.')
+    ap.add_argument('-v', '--verbose', action='store_true',
+                    help='Show progress and details.')
 
-    sys.exit(main(ap.parse_args()))
+    return ap
+
+
+if __name__ == '__main__':
+    sys.exit(main(argparser().parse_args()))